fix: Resolve clippy warnings for Rust 1.92
- Add #[allow(dead_code)] to unused test helper functions and struct fields
- Fix snake_case naming violations (refW/refH → ref_w/ref_h)
- Replace unwrap() after is_ok() with if let Ok()
- Remove needless borrows for generic args
- Fix unused imports and useless as_ref() calls
- Add missing use std::time::Instant in parallel_bench example
- Apply various auto-fixable clippy suggestions

+++ jpegli-rs/src/huffman/types.rs
@@ -437,7 +437,7 @@ mod tests {
         freq.add(17, 3000); // Run=1, size=1
         freq.add(33, 2000); // Run=2, size=1
         for i in 2..16 {
-            freq.add(i, (1000 / (i as u64 + 1)) as u64);
+            freq.add(i, 1000 / (i as u64 + 1));
         }
 
         let (mozjpeg, jpegli, moz_cost, jpg_cost) = compare_algorithms(&freq).unwrap();
